Fixed inconsistencies in the simulation caching, and split off the tile manager.
